+++ config/test.exs
@@ -5,6 +5,10 @@ config :plug_pagecache,
     agent: [
       adapter: Plug.PageCache.Adapter.Agent
     ],
+    agent_expire: [
+      adapter:     Plug.PageCache.Adapter.Agent,
+      auto_expire: 1
+    ],
     ets: [
       adapter: Plug.PageCache.Adapter.ETS,
       table:   :plug_pagecache_foo