+++ lib/plug/pagecache/adapter/agent.ex
@@ -8,7 +8,10 @@ defmodule Plug.PageCache.Adapter.Agent do
 
   # Adapter lifecycle
 
-  def init(_opts), do: { :ok, %{ entries: %{} } }
+  def init(opts) do
+    { :ok, %{ auto_expire: opts[:auto_expire] || 0,
+              entries:     %{} }}
+  end
 
 
   # Adapter methods
@@ -18,7 +21,7 @@ defmodule Plug.PageCache.Adapter.Agent do
   end
 
   def load(state, path) do
-    { state, Map.get(state.entries, path, nil) }
+    { state, load_unexpired(path, state) }
   end
 
   def remove(state, path) do
@@ -26,6 +29,25 @@ defmodule Plug.PageCache.Adapter.Agent do
   end
 
   def save(state, path, page) do
-    { %{ state | entries: Map.put(state.entries, path, page) }, :ok }
+    entry = { page, :os.system_time(:seconds) + state.auto_expire }
+
+    { %{ state | entries: Map.put(state.entries, path, entry) }, :ok }
+  end
+
+
+  # Internal methods
+
+  defp load_unexpired(path, %{ auto_expire: 0 } = state) do
+    { page, _ } = Map.get(state.entries, path, { nil, 0 })
+    page
+  end
+
+  defp load_unexpired(path, state) do
+    { page, expire_after } = Map.get(state.entries, path, { nil, 0 })
+
+    case :os.system_time(:seconds) <= expire_after do
+      true  -> page
+      false -> nil
+    end
   end
 end